Issue #5 / Sprint 7 - Add TODO list and parse disciplines CSV file
