+++ app/views/disciplines/index.html.erb
@@ -61,7 +61,7 @@
                         </div>
                         <div class="col-sm-8">
                             <div class="pull-right">
-                                <%= will_paginate @disciplines, renderer: BootstrapPagination::Rails, :previous_label => '<i class="fa fa-long-arrow-left"></i>', :next_label => '<i class="fa fa-long-arrow-right"></i>' %>
+                                <%= will_paginate @disciplines, renderer: BootstrapPagination::Rails, inner_window: 1, outer_window: 1, :previous_label => '<i class="fa fa-long-arrow-left"></i>', :next_label => '<i class="fa fa-long-arrow-right"></i>' %>
                             </div>
                         </div>
                     </div>
