--- db/migrate/20170919140927_create_disciplines.rb
@@ -1,11 +1,0 @@
-class CreateDisciplines < ActiveRecord::Migration[5.0]
-  def change
-    create_table :disciplines do |t|
-      t.string :name
-      t.string :discipline_code
-      t.references :department, foreign_key: true
-
-      t.timestamps
-    end
-  end
-end
